RAG_Real_life_use_case_Week3_Day5/Divyansh_Assignment_14_Tesla_Supply_Chain_Intelligence/requirements.txt
streamlit
langchain
langchain-community
langchain-groq
chromadb
sentence-transformers
pymupdf4llm
python-dotenv
PyJWT
rank_bm25